refactor: uffd: move peer creds function to common
Move `get_peer_process_credentials` to common.rs to make it
visible outside of the crate and use it in the malicious
demo.
